requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
pillow>=9.0.0

# Data processing
numpy>=1.23.0
pandas>=1.5.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Metrics
scikit-image>=0.19.0
scikit-learn>=1.0.0
lpips>=0.1.4

# Utilities
tqdm>=4.65.0
